requirements.txt
arrow==1.3.0
asgiref==3.7.2
Django==4.2.7
djangorestframework==3.14.0
Jinja2==3.1.2
jinja2-time==0.2.0
make==0.1.6.post2
MarkupSafe==2.1.3
Pillow==10.1.0
psycopg==3.1.13
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
